RimWorld Alpha 2 Version (Beta 0.5)
Updated all tag based vars, Now works with Alpha 2 save file.

Added Errorlog ("RwEditor_Output.txt")

Backup system now has a backup/restore button under file *NOTE Restore
not 100%*

Colonist editor now allows healing of all Pawn s with one button

Started transfering some classes postioning code to use AppPostioning

Removed ClearWeapons code no longer needing feature

DataHandler now has more Javadocs attached *Rest program todo :P* Extra
methods added to DataHandler (setData, getDataList, getFiles) Allows use
of multiple file editing based on code selection.

Start of ModLoader class has been implemented need to read Mod Files
now..

Cleaned up some code

+++ src/me/keensta/actionlisteners/resources/RSelectionListener.java
@@ -18,7 +18,7 @@ public class RSelectionListener implements ActionListener{
     @Override
     public void actionPerformed(ActionEvent e) {
         WebComboBox cb = (WebComboBox) e.getSource();
-        int index = (int) (cb.getSelectedIndex());
+        int index = cb.getSelectedIndex();
         String[] data = app.getRes().getData();
        
         if(index <= 0 || data.length == 1)
